Prevent arithmetic overflow on division
Avoid crashing the server when $minint / -1 is evaluated.

There is a much deeper problem here which is that all arithmetic
operations on signed integers are undefined in C when overflow occurs,
yet this is never checked for. Instead the prevailing assumption is that
the implementation will perform modulo arithmetic on signed integers in
a manner similar to that of unsigned integers.

A more thorough patch will be needed to correct the rest of these
infelicities.